requirements.txt
# Core RAG dependencies
faiss-cpu>=1.11.0.post1
langchain>=0.3.0
langchain-community>=0.3.0  
langchain-openai>=0.3.0
langgraph>=0.5.0

# Document processing
pdfplumber>=0.11.0
unstructured>=0.18.0
python-dotenv>=1.0.0

# Web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Type annotations (for Python < 3.11 compatibility)
typing-extensions>=4.0.0
